+++ golang-backend/internal/services/noteservice/note.go
@@ -3,10 +3,13 @@ package noteservice
 import (
 	"context"
 	"fmt"
+	"time"
 
 	"github.com/aportela/doneo/internal/database"
 	"github.com/aportela/doneo/internal/domain"
+	"github.com/aportela/doneo/internal/middlewares"
 	"github.com/aportela/doneo/internal/repositories/noterepository"
+	"github.com/aportela/doneo/internal/repositories/projecthistoryrepository"
 )
 
 type NoteService interface {
@@ -26,15 +29,78 @@ func NewService(database database.Database, repository noterepository.NoteReposi
 }
 
 func (service *noteService) AddProjectNote(ctx context.Context, projectId string, note domain.Note) error {
-	return service.repository.AddProjectNote(ctx, projectId, note)
+	tx, err := service.database.Begin()
+	if err != nil {
+		return err
+	}
+	defer func() {
+		if p := recover(); p != nil {
+			_ = tx.Rollback()
+			panic(p)
+		} else if err != nil {
+			_ = tx.Rollback()
+		}
+	}()
+	currentUserId, ok := middlewares.GetUserIDFromContext(ctx)
+	if !ok {
+		return fmt.Errorf("user ID not found in context")
+	}
+	err = service.repository.AddProjectNote(ctx, projectId, note)
+	if err != nil {
+		return err
+	}
+	err = projecthistoryrepository.NewRepository(service.database).Add(ctx, projectId, domain.ProjectHistoryOperation{CreatedBy: domain.UserBase{ID: currentUserId}, CreatedAt: time.Now(), OperationType: domain.EventProjectNoteAdded})
+	return tx.Commit()
 }
 
 func (service *noteService) UpdateProjectNote(ctx context.Context, projectId string, note domain.Note) error {
-	return service.repository.UpdateProjectNote(ctx, projectId, note)
+	tx, err := service.database.Begin()
+	if err != nil {
+		return err
+	}
+	defer func() {
+		if p := recover(); p != nil {
+			_ = tx.Rollback()
+			panic(p)
+		} else if err != nil {
+			_ = tx.Rollback()
+		}
+	}()
+	currentUserId, ok := middlewares.GetUserIDFromContext(ctx)
+	if !ok {
+		return fmt.Errorf("user ID not found in context")
+	}
+	err = service.repository.UpdateProjectNote(ctx, projectId, note)
+	if err != nil {
+		return err
+	}
+	err = projecthistoryrepository.NewRepository(service.database).Add(ctx, projectId, domain.ProjectHistoryOperation{CreatedBy: domain.UserBase{ID: currentUserId}, CreatedAt: time.Now(), OperationType: domain.EventProjectNoteUpdated})
+	return tx.Commit()
 }
 
 func (service *noteService) DeleteProjectNote(ctx context.Context, projectId string, noteId string) error {
-	return service.repository.DeleteProjectNote(ctx, projectId, noteId)
+	tx, err := service.database.Begin()
+	if err != nil {
+		return err
+	}
+	defer func() {
+		if p := recover(); p != nil {
+			_ = tx.Rollback()
+			panic(p)
+		} else if err != nil {
+			_ = tx.Rollback()
+		}
+	}()
+	currentUserId, ok := middlewares.GetUserIDFromContext(ctx)
+	if !ok {
+		return fmt.Errorf("user ID not found in context")
+	}
+	err = service.repository.DeleteProjectNote(ctx, projectId, noteId)
+	if err != nil {
+		return err
+	}
+	err = projecthistoryrepository.NewRepository(service.database).Add(ctx, projectId, domain.ProjectHistoryOperation{CreatedBy: domain.UserBase{ID: currentUserId}, CreatedAt: time.Now(), OperationType: domain.EventProjectNoteDeleted})
+	return tx.Commit()
 }
 
 func (service *noteService) GetProjectNotes(ctx context.Context, projectId string) ([]domain.Note, error) {
